+++ Pandora/Paradigm/Primary/Functor/Predicate.hs
@@ -4,7 +4,8 @@ import Pandora.Core.Morphism ((!))
 import Pandora.Pattern.Category ((.), ($))
 import Pandora.Pattern.Functor.Contravariant (Contravariant ((>$<)))
 import Pandora.Pattern.Functor.Determinable (Determinable (determine))
-import Pandora.Paradigm.Primary.Object.Boolean (Boolean (True))
+import Pandora.Paradigm.Primary.Functor.Maybe (Maybe (Just, Nothing))
+import Pandora.Paradigm.Primary.Object.Boolean (Boolean (True), (?))
 
 newtype Predicate a = Predicate (a -> Boolean)
 
@@ -13,3 +14,6 @@ instance Contravariant Predicate where
 
 instance Determinable Predicate where
 	determine = Predicate (True !)
+
+satisfy :: Predicate a -> a -> Maybe a
+satisfy (Predicate p) x = p x ? Just x $ Nothing